+++ lms/djangoapps/courseware/tests/test_draft_modulestore.py
@@ -4,7 +4,7 @@ from django.test.utils import override_settings
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore import Location
 
-from modulestore_config import TEST_DATA_DRAFT_MONGO_MODULESTORE
+from course_groups.tests.modulestore_config import TEST_DATA_DRAFT_MONGO_MODULESTORE
 
 
 @override_settings(MODULESTORE=TEST_DATA_DRAFT_MONGO_MODULESTORE)
