+++ lms/djangoapps/instructor/tests/test_download_csv.py
@@ -17,7 +17,7 @@ from django.core.urlresolvers import reverse
 
 from courseware.access import _course_staff_group_name
 from courseware.tests.helpers import LoginEnrollmentTestCase
-from courseware.tests.modulestore_config import TEST_DATA_XML_MODULESTORE
+from course_groups.tests.modulestore_config import TEST_DATA_XML_MODULESTORE
 from xmodule.modulestore.django import modulestore
 import xmodule.modulestore.django
